+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v8.5.0 (2025-12-08)
+
+- Adds the following functions:
+  - `fedex.verifyBillingAddress`
+  - `fedex.generatePin`
+  - `fedex.validatePin`
+  - `fedex.validateInvoice`
+- Adds `details` property of a `Rate` object
+
 ## v8.4.1 (2025-12-01)
 
 - Adds missing `apiKeys` field to `BaseUser`